requirements.txt
streamlit
groq
python-dotenv
streamlit-mic-recorder